rename mspn_head to msmu_head
