+++ mmpose/models/keypoint_heads/__init__.py
@@ -1,10 +1,9 @@
 from .bottom_up_higher_resolution_head import BottomUpHigherResolutionHead
 from .bottom_up_simple_head import BottomUpSimpleHead
-from .top_down_mspn_head import TopDownMSPNHead
-from .top_down_multi_stage_head import TopDownMultiStageHead
+from .top_down_multi_stage_head import TopDownMSMUHead, TopDownMultiStageHead
 from .top_down_simple_head import TopDownSimpleHead
 
 __all__ = [
-    'TopDownSimpleHead', 'TopDownMultiStageHead', 'TopDownMSPNHead',
+    'TopDownSimpleHead', 'TopDownMultiStageHead', 'TopDownMSMUHead',
     'BottomUpHigherResolutionHead', 'BottomUpSimpleHead'
 ]
